test(pallet-market-benchmarks): Move test-only stuff from mock to test, and fix deps
