+++ pallets/market/benchmarks/src/benchmarking.rs
@@ -486,7 +486,7 @@ mod benchmarks {
 
     impl_benchmark_test_suite! {
         Pallet,
-        crate::mock::new_test_ext(),
+        crate::test::new_test_ext(),
         crate::mock::Test,
     }
 }
