+++ pallets/market/benchmarks/src/lib.rs
@@ -5,8 +5,8 @@ extern crate alloc;
 #[cfg(feature = "runtime-benchmarks")]
 pub mod benchmarking;
 
-#[cfg(any(feature = "runtime-benchmarks", test))]
-pub(crate) mod mock;
+#[cfg(test)]
+mod mock;
 
 #[cfg(any(feature = "runtime-benchmarks", test))]
 mod pallet;
